fix：修复 Group 查询不支持 aot 问题 perf：优化非 aot 下分组查询性能（减少反射）
